feature: add global build constant values
In the same way an android application can access the:
  - packageName(com.example.appName)
  - appName(appName)
  - version(context.getPackageManager().getPackageInfo(context.getPackageName(),
  0).versionName)

Add global values that can retrieve those informations,
this is necessary for the making of the (flutter maintained)
[package_info](https://github.com/flutter/plugins/tree/master/packages/package_info)
flutter plugin.

This will also came useful for the future `hover init-plugin` command,
as the default `flutter create --template=plugin plugin_name` creates a
template plugin that retrieve the `android.os.Build.VERSION.RELEASE`
constant on android.  The mapping value for go-flutter is
`flutter.PlatformVersion()` (go-flutter release used for the project)